Move all route registering to lib/routes
