+++ lib/routes.js
@@ -10,11 +10,52 @@ const index = require('./controllers/index');
 const doc = require('./controllers/document');
 const series = require('./controllers/series');
 const config = require('./config');
+const indexing = require('../controllers/indexing');
+const users = require('../controllers/users');
+const keystonePlugin = require('../plugins/keystone');
+const authPlugin = require('../plugins/auth');
 
 /**
  * Application routes
  */
 module.exports = function(app) {
+  app.post('/index/asset', indexing.asset);
+  app.get('/healthz', (req, res) => { res.send('ok'); });
+  // Setup an API endpoint for fetching contributions. We expect a asset type
+  // and a numeric page-number.
+  app.get('/min-side/contributions/:assetType/:pageNo(\\d+)', users.fetchUserContributions);
+
+  keystonePlugin.registerRoutes(app);
+  authPlugin.registerRoutes(app);
+
+  // Indexing endpoints
+  app.get('/index/recent', (req, res) => {
+    if (config.kbhAccessKey && req.query.accesskey !== config.kbhAccessKey) {
+      res.status(401);
+      return res.send('Accesskey required.');
+    }
+    const state = {
+      mode: 'recent',
+      reference: req.query.timeframe
+    };
+
+    indexing(state);
+    return res.send('Indexing of recent assets started.');
+  });
+
+  app.get('/index/all', (req, res) => {
+    if (config.kbhAccessKey && req.query.accesskey !== config.kbhAccessKey) {
+      res.status(401);
+      return res.send('Accesskey required.');
+    }
+    const state = {
+      mode: 'all',
+    };
+
+    indexing(state);
+    return res.send('Full reindexing started.');
+  });
+
   app.use('/was',
     (req, res) => res.redirect('https://www.was.digst.dk/kbhbilleder-dk')
   );
